requirements.txt
databricks-sql-connector~=3.4.0
databricks-sdk~=0.33.0
gradio~=4.44.0
huggingface-hub~=0.35.3
pandas~=2.2.3
plotly~=5.22.0
strands-agents
python-dotenv~=1.0.0
pyyaml~=6.0.1
sqlparse~=0.5.0
numpy~=1.26.4
